Updates after the second round of review
